chore: refactoring v3 macro structure
