+++ scripts/generateSDK/generate.js
@@ -77,11 +77,7 @@ async function generateSdk({ url, sdkName }) {
 const generateSdkInBatch = () => {
   const config = [
     {
-      sdkName: 'GatewayV2',
-      url: 'https://develop.protocol.mygateway.xyz/graphql',
-    },
-    {
-      sdkName: 'GatewayV3',
+      sdkName: 'Gateway',
       url: 'https://v3-dev.protocol.mygateway.xyz/graphql',
     },
   ];
