+++ src/modules/request/request.ts
@@ -6,8 +6,11 @@ import {
   requestsReceived_queryQueryVariables,
   requestsSent_queryQueryVariables,
 } from '../../../gatewaySdk/sources/GatewayV3';
-import { errorHandler } from '../../utils/errorHandler';
-import { isUUIDValid, validateObjectProperties } from '../../utils/validators';
+import { errorHandler } from '../../utils/error-handler';
+import {
+  isUUIDValid,
+  validateObjectProperties,
+} from '../../utils/validator-service';
 
 export class Request {
   public sdk: Sdk;
