+++ src/types/types.ts
@@ -1,17 +1,9 @@
-export enum UserIdentifierType {
-  EMAIL = 'EMAIL',
-  EVM = 'EVM',
-  GATEWAY_ID = 'GATEWAY_ID',
-  SOLANA = 'SOLANA',
-  USER_ID = 'USER_ID',
-}
-
 export enum SignCipherEnum {
   ED25519 = 'ED25519',
   SECP256K1 = 'SECP256K1',
 }
 
-export enum UserIdentifierTypeV3 {
+export enum UserIdentifierType {
   EMAIL = 'EMAIL',
   EVM = 'EVM',
   POK = 'POKT',
@@ -28,11 +20,6 @@ export enum PDAStatus {
 }
 
 export enum OrganizationIdentifierType {
-  GATEWAY_ID = 'GATEWAY_ID',
-  ORG_ID = 'ORG_ID',
-}
-
-export enum OrganizationIdentifierTypeV3 {
   ORG_DID = 'ORG_DID',
   ORG_ID = 'ORG_ID',
   USERNAME = 'USERNAME',
